ai-engineer-bootcamp/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-engineer-bootcamp"
version = "0.1.0"
description = "Base project for Gemini API integration"
requires-python = ">=3.10"
dependencies = [
  "google-genai",
  "python-dotenv",
  "rich",
]

[project.scripts]
ai-engineer-bootcamp = "main:main"
